+++ src/core/element_container/ConverterStationContainer.hpp
@@ -168,13 +168,15 @@ class LS2G_API ConverterStationContainer final : public OneSideContainer_PQ, pub
         void init_q_vector(int nb_bus,
                            Eigen::Ref<Eigen::VectorXi> total_gen_per_bus,
                            Eigen::Ref<RealVect> total_q_min_per_bus,
-                           Eigen::Ref<RealVect> total_q_max_per_bus) const;
+                           Eigen::Ref<RealVect> total_q_max_per_bus,
+                           const std::vector<bool> & solved_by_algo) const;
         void set_q(const Eigen::Ref<const RealVect> & reactive_mismatch,
                    const SolverBusIdVect & id_grid_to_solver,
                    bool ac,
                    const Eigen::Ref<const Eigen::VectorXi> & total_gen_per_bus,
                    const Eigen::Ref<const RealVect> & total_q_min_per_bus,
-                   const Eigen::Ref<const RealVect> & total_q_max_per_bus);
+                   const Eigen::Ref<const RealVect> & total_q_max_per_bus,
+                   const std::vector<bool> & solved_by_algo);
         void get_vm_for_dc(Eigen::Ref<RealVect> Vm);
         void set_vm(Eigen::Ref<CplxVect> V, const SolverBusIdVect & id_grid_to_solver) const;